runAllTestsInFolder to TestSuite.cpp
